backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic[email]
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
numpy
pandas
databases
gunicorn
psycopg2-binary==2.9.9
bcrypt<4.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
requests==2.31.0